Update Chart.yaml to v2 Helm api
The requirements.yaml file is no more

Here, we also refer to a new controller version v2.23.0 which we
described before. We will actually need it published somewhere so we can
use it for testing. (Creating the hephy.pro ChartMuseum for this now.)

Without a published controller chart that includes the v1 ingress,
Workflow is not installable with Experimental Native Ingress support
onto clusters of v1.22 Kubernetes releases or higher. Since I am the
only person I know that uses Experimental Native Ingress support in
Hephy Workflow, it is a priority (for me) to release this ASAP!

Let's do it as a beta version, we'll publish it for ourselves and make a
great show of testing the release in public where you can replicate the
results and experiment for your own proof before Hephy Workflow v2.24 is
published, (or Controller chart v2.23.0, which is prerequisite to that.)